Refactoring client index for woo-ai plugin
