+++ plugins/woo-ai/src/product-description/index.ts
@@ -1,1 +1,12 @@
-export * from './product-description-button-container';
+/**
+ * Internal dependencies
+ */
+import { renderWrappedComponent } from '../utils';
+import { WriteItForMeButtonContainer } from './product-description-button-container';
+
+if ( window.JP_CONNECTION_INITIAL_STATE?.connectionStatus?.isActive ) {
+	renderWrappedComponent(
+		WriteItForMeButtonContainer,
+		document.getElementById( 'woocommerce-ai-app-product-gpt-button' )
+	);
+}
