+++ plugins/woo-ai/src/utils/index.ts
@@ -6,3 +6,4 @@ export * from './tiny-tools';
 export * from './productDataInstructionsGenerator';
 export * from './categorySelector';
 export * from './htmlEntities';
+export * from './renderWrappedComponent';
